requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
pydantic>=2.8,<3.0
pydantic-settings>=2.3,<3.0
httpx>=0.27,<1.0
pytest>=8.3,<9.0
black>=24.3,<25.0
ruff>=0.5,<1.0
Pillow>=10.4,<11.0
